requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
openai==2.33.0
scikit-learn==1.8.0
pandas==3.0.2
numpy==2.2.6
python-multipart==0.0.20
python-dotenv==1.2.1
